asset-clustering/training-tuning-inference/requirements.txt
numpy
cudf-cu11
cuml-cu11
cupy
pandas
scikit-learn
click

